Change background spinner and set spinner as util
