+++ Cosmetic/Assets.xcassets/ColorSet/main-font-color.colorset/Contents.json
@@ -1,0 +1,38 @@
+{
+  "info" : {
+    "version" : 1,
+    "author" : "xcode"
+  },
+  "colors" : [
+    {
+      "idiom" : "universal",
+      "color" : {
+        "color-space" : "srgb",
+        "components" : {
+          "red" : "0.868",
+          "alpha" : "1.000",
+          "blue" : "0.745",
+          "green" : "0.252"
+        }
+      }
+    },
+    {
+      "idiom" : "universal",
+      "appearances" : [
+        {
+          "appearance" : "luminosity",
+          "value" : "dark"
+        }
+      ],
+      "color" : {
+        "color-space" : "srgb",
+        "components" : {
+          "red" : "0.962",
+          "alpha" : "1.000",
+          "blue" : "0.837",
+          "green" : "0.613"
+        }
+      }
+    }
+  ]
+}
