+++ Cosmetic/Library.swift
@@ -1,0 +1,41 @@
+//
+//  Util.swift
+//  Cosmetic
+//
+//  Created by Omp on 23/11/2562 BE.
+//  Copyright © 2562 Omp. All rights reserved.
+//
+
+import UIKit
+
+class Library {
+    
+}
+
+//MARK: - Spinner
+var spinView: UIView!
+extension UIViewController{
+    
+    func showSpinner(onView :UIView) {
+        
+        let spinnerView = UIView.init(frame: onView.bounds)
+        spinnerView.backgroundColor = UIColor.init(named: "spinner-bg")
+        let ai = UIActivityIndicatorView.init(style: .whiteLarge)
+        ai.startAnimating()
+        ai.center = spinnerView.center
+        
+        DispatchQueue.main.async {
+            spinnerView.addSubview(ai)
+            onView.addSubview(spinnerView)
+        }
+        
+        spinView = spinnerView
+    }
+    
+    func removeSpinner() {
+        DispatchQueue.main.async {
+            spinView?.removeFromSuperview()
+            spinView = nil
+        }
+    }
+}
