+++ Cosmetic/SearchTableViewController.swift
@@ -108,29 +108,3 @@ extension SearchTableViewController: UISearchBarDelegate{
         navigationController?.pushViewController(searchDetailVc, animated: true)
     }
 }
-
-var vSpinner :UIView?
-extension SearchTableViewController{
-
-    func showSpinner(onView :UIView) {
-        let spinnerView = UIView.init(frame: onView.bounds)
-        spinnerView.backgroundColor = UIColor.init(red: 0.5, green: 0.5, blue: 0.5, alpha: 0.5)
-        let ai = UIActivityIndicatorView.init(style: .whiteLarge)
-        ai.startAnimating()
-        ai.center = spinnerView.center
-        
-        DispatchQueue.main.async {
-            spinnerView.addSubview(ai)
-            onView.addSubview(spinnerView)
-        }
-        
-        vSpinner = spinnerView
-    }
-    
-    func removeSpinner() {
-        DispatchQueue.main.async {
-            vSpinner?.removeFromSuperview()
-            vSpinner = nil
-        }
-    }
-}
